kutuphane-sistem/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.5
beautifulsoup4==4.12.2
lxml==4.9.3
python-multipart==0.0.6
pydantic==2.5.2
sickle==0.7.0
langchain==0.1.9
langchain-openai==0.0.5 